pyproject.toml
[project]
name = "helix"
version = "4.0.0"
description = "HELIX v4 - AI Development Platform"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "httpx>=0.24",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.scripts]
helix = "helix.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
